# KIKIMR-19522 BTreeIndex Iter Remove single run optimization