Seek benchmarks `yag test -r -F benchmark::TPartIndexIteratorFixture/DoReads/*/0/0/*/*/1`

(`TPartIndexIteratorFixture/DoReads/b-tree/groups=0/history=0/reverse/ESeek/items=1`)

Before:

```
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
TPartIndexIteratorFixture/DoReads/0/0/0/0/1/1       18.8 us         7.74 us        89449
TPartIndexIteratorFixture/DoReads/1/0/0/0/1/1       7.41 us         7.39 us        93294
TPartIndexIteratorFixture/DoReads/0/0/0/1/1/1       7.42 us         7.21 us        98974
TPartIndexIteratorFixture/DoReads/1/0/0/1/1/1       7.58 us         7.57 us        93598
TPartIndexIteratorFixture/DoReads/0/0/0/0/2/1       8.22 us         6.94 us       104775
TPartIndexIteratorFixture/DoReads/1/0/0/0/2/1       8.43 us         7.26 us        96510
TPartIndexIteratorFixture/DoReads/0/0/0/1/2/1       12.5 us         7.20 us       105019
TPartIndexIteratorFixture/DoReads/1/0/0/1/2/1       7.14 us         7.13 us        96782
TPartIndexIteratorFixture/DoReads/0/0/0/0/3/1       13.6 us         7.58 us        88728
TPartIndexIteratorFixture/DoReads/1/0/0/0/3/1       9.47 us         7.64 us        93685
TPartIndexIteratorFixture/DoReads/0/0/0/1/3/1       9.68 us         6.96 us       102325
TPartIndexIteratorFixture/DoReads/1/0/0/1/3/1       11.0 us         7.05 us        97635
```

After:

```
----------------------------------------------------------------------------------------
Benchmark                                              Time             CPU   Iterations
----------------------------------------------------------------------------------------
TPartIndexIteratorFixture/DoReads/0/0/0/0/1/1       11.4 us         7.37 us        94450
TPartIndexIteratorFixture/DoReads/1/0/0/0/1/1       9.49 us         7.77 us        92168
TPartIndexIteratorFixture/DoReads/0/0/0/1/1/1       7.87 us         7.18 us        99113
TPartIndexIteratorFixture/DoReads/1/0/0/1/1/1       9.71 us         7.69 us        86517
TPartIndexIteratorFixture/DoReads/0/0/0/0/2/1       10.2 us         7.37 us        90860
TPartIndexIteratorFixture/DoReads/1/0/0/0/2/1       10.3 us         7.61 us        93749
TPartIndexIteratorFixture/DoReads/0/0/0/1/2/1       7.81 us         6.87 us        96138
TPartIndexIteratorFixture/DoReads/1/0/0/1/2/1       8.95 us         7.36 us        99316
TPartIndexIteratorFixture/DoReads/0/0/0/0/3/1       9.64 us         7.18 us        96818
TPartIndexIteratorFixture/DoReads/1/0/0/0/3/1       11.0 us         7.77 us        95898
TPartIndexIteratorFixture/DoReads/0/0/0/1/3/1       15.1 us         7.36 us       103496
TPartIndexIteratorFixture/DoReads/1/0/0/1/3/1       7.41 us         7.37 us        94385
```

No differences noticed